Move PhotoFieldViewModel into editobservation package
