+++ gnd/src/main/java/com/google/android/gnd/ui/common/ViewModelModule.java
@@ -19,8 +19,8 @@ package com.google.android.gnd.ui.common;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 import com.google.android.gnd.MainViewModel;
-import com.google.android.gnd.ui.common.photoview.PhotoFieldViewModel;
 import com.google.android.gnd.ui.editobservation.EditObservationViewModel;
+import com.google.android.gnd.ui.editobservation.PhotoFieldViewModel;
 import com.google.android.gnd.ui.home.HomeScreenViewModel;
 import com.google.android.gnd.ui.home.featuresheet.FeatureSheetViewModel;
 import com.google.android.gnd.ui.home.featuresheet.ObservationListViewModel;
