+++ gnd/src/main/res/layout/photo_input_field.xml
@@ -28,7 +28,7 @@
       type="com.google.android.gnd.ui.editobservation.EditObservationFragment" />
     <variable
       name="photoFieldViewModel"
-      type="com.google.android.gnd.ui.common.photoview.PhotoFieldViewModel" />
+      type="com.google.android.gnd.ui.editobservation.PhotoFieldViewModel" />
   </data>
 
   <LinearLayout
